moved build_model_and_load_weights into torch_utils
