+++ src/ednaml/utils/__init__.py
@@ -256,44 +256,6 @@ model_weights = {
 
 from ednaml.utils.LabelMetadata import LabelMetadata
 
-from ednaml.models.ModelAbstract import ModelAbstract
-def build_model_and_load_weights(
-    config_file: str,
-    model_class: Type[ModelAbstract] = None,
-    epoch: int = 0,
-    custom_metadata: LabelMetadata = None,
-    add_filehandler: bool = False,
-    add_streamhandler: bool = True,
-):
-    """Generates a model using a configuration file, and loads a specific saved epoch
-
-    Args:
-        config_file (str): _description_
-        model_class (Type[ModelAbstract], optional): _description_. Defaults to None.
-        epoch (int, optional): _description_. Defaults to 0.
-
-    Returns:
-        _type_: _description_
-    """
-    from ednaml.core import EdnaML
-
-    eml = EdnaML(
-        config=config_file,
-        mode="test",
-        test_only=True,
-        add_filehandler=add_filehandler,
-        add_streamhandler=add_streamhandler,
-    )
-    eml.labelMetadata = custom_metadata  # TODO this needs to be fixed with the actual label metadata...or tell users to not infer things :|
-    if model_class is not None:
-        eml.addModelClass(model_class=model_class)
-    eml.buildModel()
-    if epoch is None:
-        # means we get the most recent epoch available...
-        epoch = eml.getPreviousStop()
-    eml.loadEpoch(epoch=epoch)
-    return eml.model
-
 
 def merge_dictionary_on_key_with_copy(a, b, path=None):
     from copy import deepcopy
